re-organization with steps hooklib
This is just a reorg for the hooklib stuff the spells can rely on.
