--- share/hooklib/README.md
@@ -1,2 +1,0 @@
-# hooklib
-> common shell functions for packaging processing tasks
